+++ GradesApp/Person.cs
@@ -4,7 +4,6 @@
     {
         public string FirstName { get; set; }
         public string LastName { get; set; }
-
         public Person(string firstName, string lastName)
         {
             this.FirstName = firstName;
