+++ GradesApp/SavedStudent.cs
@@ -6,7 +6,7 @@ using System.Text;
 
 namespace GradesApp
 {
-    public class SavedStudent : StudentBase
+    public class SavedStudent : AddingGrades
     {
         private const string fileName = "_grades.txt";
         private string fullFileName;
@@ -28,56 +28,6 @@ namespace GradesApp
                     {
                         CheckIsGradeUnder3();
                     }
-                }
-            }
-        }
-
-        public override void AddGrade(string grade)
-        {
-            double convertedGradeToDouble = char.GetNumericValue(grade[0]);
-            if (grade.Length == 2 && char.IsDigit(grade[0]) && grade[0] <= '6' && (grade[1] == '+' || grade[1] == '-'))
-            {
-                switch (grade[1])
-                {
-                    case '+':
-                        double gradePlus = convertedGradeToDouble + 0.50;
-                        if (gradePlus > 1 && gradePlus <= 6)
-                        {
-                            AddGrade(gradePlus);
-                        }
-                        else
-                        {
-                            throw new ArgumentException($"Invalid argument: {nameof(grade)}. Only grades from 1 to 6 are allowed!");
-                        }
-                        break;
-
-                    case '-':
-                        double gradeMinus = convertedGradeToDouble - 0.25;
-                        if (gradeMinus > 1 && gradeMinus <= 6)
-                        {
-                            AddGrade(gradeMinus);
-                        }
-                        else
-                        {
-                            throw new ArgumentException($"Invalid argument: {nameof(grade)}. Only grades from 1 to 6 are allowed!");
-                        }
-                        break;
-
-                    default:
-                        throw new ArgumentException($"Invalid argument: {nameof(grade)}. Only grades from 1 to 6 are allowed!");
-                }
-            }
-            else
-            {
-                double gradeDouble = 0;
-                var isParsed = double.TryParse(grade, out gradeDouble);
-                if (isParsed && gradeDouble > 0 && gradeDouble <= 6)
-                {
-                    AddGrade(gradeDouble);
-                }
-                else
-                {
-                    throw new ArgumentException($"Invalid argument: {nameof(grade)}. Only grades from 1 to 6 are allowed!");
                 }
             }
         }
@@ -114,6 +64,5 @@ namespace GradesApp
                 }
             }
         }
-
     }
 }
